1. exceptions.py: Added base_message to exceptions and init for message. 2. SourceAdminForm: data >> cleaned_data. 3. DRY in methods.py in handle_response. 4. Converter model value FloatField >> DecimalField. 5. tasks.py: some changes to clean code with exceptions. 6. JSONEncoder: added Decimal case.

+++ currency/converter/models.py
@@ -76,7 +76,9 @@ class Converter(models.Model):
         verbose_name='Output ticker',
         help_text='Base currency that is using as reference in the source',
     )
-    value = models.FloatField(
+    value = models.DecimalField(
+        max_digits=10,
+        decimal_places=4,
         verbose_name='Value',
     )
     coefficient = models.IntegerField(
